feat: add vectorize mod

+++ src/config.rs
@@ -1,9 +1,9 @@
 pub const MAX_AMOUNT: f32 = 10_000.0;
-pub const MAX_INSTALLMENTS: u8 = 12;
+pub const MAX_INSTALLMENTS: f32 = 12.0;
 pub const AMOUNT_VS_AVG_RATIO: f32 = 10.0;
 pub const MAX_MINUTES: f32 = 1440.0;
 pub const MAX_KM: f32 = 1000.0;
-pub const MAX_TX_COUNT_24H: u8 = 20;
+pub const MAX_TX_COUNT_24H: f32 = 20.0;
 pub const MAX_MERCHANT_AVG_AMOUNT: f32 = 10_000.0;
 
 pub fn mcc_risk(mcc: &str) -> f32 {
